Reporting about removed and changed objects.

+++ scripts/generateUpdatePosts.sh
@@ -71,11 +71,42 @@ if [[ $x -ge $latestPostVersion ]] || [ $x == "Start" ]; then
 			
 			newObjCount=$((newObjCount+1))
 
-		done < <(hg status --rev OneLife_v$x:OneLife_v$newerVersion objects/ | grep "A " | sed 's/A //')
+		done < <(hg status --rev OneLife_v$x:OneLife_v$newerVersion objects/ -X objects/nextObjectNumber.txt | grep "A " | sed 's/A //')
 
-		echo "$newObjCount objects in report for v$newerVersion"
+		echo "$newObjCount new objects in report for v$newerVersion"
 
-		if [[ $newObjCount -eq 0 ]];
+
+		changedObjCount=0
+		while read y;
+		do
+			objName=$(cat $y | sed -n 2p )
+			
+			echo "(changed) $objName<br>" >> $reportFile
+			
+			changedObjCount=$((changedObjCount+1))
+
+		done < <(hg status --rev OneLife_v$x:OneLife_v$newerVersion objects/-X objects/nextObjectNumber.txt | grep "M " | sed 's/M //')
+
+		echo "$changedObjCount changed objects in report for v$newerVersion"
+
+
+
+		removedObjCount=0
+		while read y;
+		do
+			objName=$(cat $y | sed -n 2p )
+			
+			echo "(removed) $objName<br>" >> $reportFile
+			
+			removedObjCount=$((removedObjCount+1))
+
+		done < <(hg status --rev OneLife_v$x:OneLife_v$newerVersion objects/-X objects/nextObjectNumber.txt | grep "R " | sed 's/R //')
+
+		echo "$removedObjCount removed objects in report for v$newerVersion"
+
+
+
+		if [[ $newObjCount -eq 0 && $changedObjCount -eq 0 && $removedObjCount =eq 0 ]];
 		then
 			rm $reportFile
 		fi
